+++ Wable-iOS/App/SceneDelegate.swift
@@ -91,7 +91,8 @@ private extension SceneDelegate {
                     loginRepository: loginRepository,
                     userSessionRepository: userSessionRepository
                 ),
-                updateUserSessionUseCase: FetchUserInformationUseCase(repository: userSessionRepository)
+                updateUserSessionUseCase: FetchUserInformationUseCase(repository: userSessionRepository),
+                userProfileUseCase: UserProfileUseCase(repository: ProfileRepositoryImpl())
             )
         )
     }
